+++ src/KMonad/Keyboard/IO/Mac/IOKitSource.hs
@@ -1,5 +1,6 @@
 module KMonad.Keyboard.IO.Mac.IOKitSource
   ( iokitSource
+  , iokitRegistryIDSource
   )
 where
 
@@ -18,7 +19,7 @@ import KMonad.Keyboard.IO.Mac.Types
 
 -- | Use the mac c-api to `grab` a keyboard
 foreign import ccall "grab_kb"
-  grab_kb :: CString -> IO Word8
+  grab_kb :: CString -> Word64 -> Word8 -> IO Word8
 
 -- | Release the keyboard hook
 foreign import ccall "release_kb"
@@ -39,7 +40,23 @@ makeLenses ''EvBuf
 iokitSource :: HasLogFunc e
   => Maybe String
   -> RIO e (Acquire KeySource)
-iokitSource name = mkKeySource (iokitOpen name) iokitClose iokitRead
+iokitSource name = iokitSourceWith name 0 0
+
+-- | Return a KeySource that seizes exactly the keyboard with this IOKit
+-- registry entry ID. Unlike 'iokitSource', this never matches a second
+-- keyboard with the same product name.
+iokitRegistryIDSource :: HasLogFunc e
+  => Word64
+  -> RIO e (Acquire KeySource)
+iokitRegistryIDSource registryID = iokitSourceWith Nothing registryID 1
+
+iokitSourceWith :: HasLogFunc e
+  => Maybe String
+  -> Word64
+  -> Word8
+  -> RIO e (Acquire KeySource)
+iokitSourceWith name registryID useRegistryID =
+  mkKeySource (iokitOpen name registryID useRegistryID) iokitClose iokitRead
 
 
 --------------------------------------------------------------------------------
@@ -47,14 +64,17 @@ iokitSource name = mkKeySource (iokitOpen name) iokitClose iokitRead
 -- | Ask IOKit to open keyboards matching the specified name
 iokitOpen :: HasLogFunc e
   => Maybe String
+  -> Word64
+  -> Word8
   -> RIO e EvBuf
-iokitOpen m = do
+iokitOpen m registryID useRegistryID = do
   logInfo "Opening IOKit devices"
   liftIO $ do
 
     case m of
-      Nothing -> void $ grab_kb nullPtr
-      Just s  -> void $ withCString s grab_kb
+      Nothing -> void $ grab_kb nullPtr registryID useRegistryID
+      Just s  -> void $ withCString s $ \product ->
+        grab_kb product registryID useRegistryID
 
     buf <- malloc @MacKeyEvent
     pure $ EvBuf buf